fix continuous replication occurring after identical normal one
Both forms of replication had the same child ID in the replication supervisor.
As a result, a continuous replication triggered after an otherwise-identical
normal replication would "forget" that it was continuous.

The tricky part about this patch is that it lets a normal and a continuous
replication with the same source and target execute simultaneously.  That's
inefficient, but other solutions are convoluted or give surprising results to
the client.


git-svn-id: http://svn.apache.org/repos/asf/couchdb/trunk@810358 13f79535-47bb-0310-9956-ffa450edef68