Fix for building CouchDB on Snow Leapord.
I was unable to track down the root cause of this bug, but for some reason, having -ljs in the link command for couch_erl_driver was causing a segfault when loading the latter in Erlang.

This patch just removes -ljs from the global LIBS variable and places it specificially in the libs for couchjs.



git-svn-id: http://svn.apache.org/repos/asf/couchdb/trunk@811910 13f79535-47bb-0310-9956-ffa450edef68