Fixes COUCHDB-504
Thanks to Bob Dionne for figuring out the underlying cause. Thanks to Benoit for making my initial patch alot cleaner.

This version is basically Benoit's patch with two extra conditions for sending the Keep-Alive header. It must be a HTTP/1.0 connection and there must not be a Connection header already defined.



git-svn-id: http://svn.apache.org/repos/asf/couchdb/trunk@816033 13f79535-47bb-0310-9956-ffa450edef68