Add ETap to CouchDB's SVN repository.
This pulls in ETap as an included dependency. As per directions of the ASF the NOTICE and LICENSE have been updated. Tests have been updated to include etap on Erlang's code path. license.skip was updated for make distcheck. This only affects running the Erlang test suite. 

Noah is awesome.



git-svn-id: http://svn.apache.org/repos/asf/couchdb/trunk@819799 13f79535-47bb-0310-9956-ffa450edef68