Fix weird error with JS_DefineFunctions call.
I found a fairly odd case with JS_DefineFunctions returning success without actually having defined any functions on the global object. This was happening on an Ubuntu 9.10 machine with Spidermonkey 1.7 which is most odd.

I refactored the code to iterate over the JSFunctionSpec array and define the functions one by one which appears to alleviate the issue.



git-svn-id: http://svn.apache.org/repos/asf/couchdb/trunk@885041 13f79535-47bb-0310-9956-ffa450edef68